Version 56
-- There were several new input files added to this revision in .cha and .ini sections;
-- Some existing input files had minor changes
-- The DTABLE.DTL file was split into three files;
-- New output files added for hru pathogen and pesticide
-- All existing output files were renamed
-- Many new subroutines were added to this revision
-- Many calibration routines were renamed

+++ source_codes/cli_initwgn.f90
@@ -112,7 +112,7 @@
                                             !              |dormant)     
       integer :: mon                        !none          |monthly counter
       real :: mdays                         !none          |number of days in the month
-      integer :: j                             !none          |counter
+      real :: j                             !none          |counter
       real :: m1                            !none          |array location (see definition of ndays)
       real ::  nda                          !julian date   |julian date of last day in the month
       real :: cli_dstn1                     !              |
